requirements.txt
# LangChain Core
langchain
langchain-core>=0.2.38


# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Environment Variable Management
python-dotenv>=1.0.0

# Machine Learning Utilities
numpy
scikit-learn

#structured output library
pydantic

#streamlit for UI in python
streamlit